third_party/blink/web_tests/fast/forms/state-restore-dynamic-controls.html

<!DOCTYPE html>
<script src="../../resources/testharness.js"></script>
<script src="../../resources/testharnessreport.js"></script>
<script src="../../external/wpt/html/interaction/focus/the-autofocus-attribute/resources/utils.js"></script>
<script src="file/resources/file-drag-common.js"></script>
<style>
/* Make the top-left corner of the iframe is equal to the top-left corner
   of the body.  It's required for dragFilesOntoInput(). */
body {
  margin: 0;
}
iframe {
  border: none;
}
</style>
<body>
<iframe src="resources/state-restore-dynamic-controls-frame.html"></iframe>
<script>
window.inputOrChangeEvents = [];

promise_test(async t => {
  await waitForEvent(window, 'load', {once:true});
  const iframe = document.querySelector('iframe');
  let doc = iframe.contentDocument;
  const container = doc.querySelector('div');
  // Change states of controls
  container.firstChild.click();
  container.lastChild.focus();
  eventSender.keyDown('z');
  dragFilesOntoInput(doc.querySelector('#file-modified'), ['foo.txt']);
  doc.querySelector('#date-modified').value = '2020-01-08';
  doc.querySelector('#textarea-modified').focus();
  eventSender.keyDown('y');
  doc.querySelector('#select-modified').focus();
  eventSender.keyDown('2');
  assert_true(container.firstChild.checked, 'sanity check for a checkbox');
  assert_equals(container.lastChild.value, 'z', 'sanity check for a text field');
  assert_equals(doc.querySelector('#textarea-modified').value, 'y', 'sanity check for a textarea');
  assert_equals(doc.querySelector('#select-modified').value, '2', 'sanity check for a select');

  // Flush asynchronous input/change events
  await timeOut(t, 0);

  // Navigate
  iframe.src = 'data:text/html,<h1></h1>';
  await waitForEvent(iframe, 'load', {once:true});

  inputOrChangeEvents = []
  // Navigate back
  history.back();
  await waitForEvent(iframe, 'load', {once:true});
  // Wait until finishing the restore task.
  await timeOut(t, 0);
  doc = iframe.contentDocument;
  const inputs = doc.querySelectorAll('input');
  assert_true(inputs[0].checked, 'Checkbox state should be restored.');
  assert_equals(inputs[1].value, 'z', 'Text field state should be restored.');
  assert_false(inputs[2].checked, 'Checkbox should have initial value.');
  assert_equals(inputs[3].value, '', 'Text field should have initial value.');
  assert_equals(doc.querySelector('textarea').value, 'y');
  assert_equals(doc.querySelector('select').value, '2');

  // Wait for asynchronous input/change events
  await timeOut(t, 0);
  assert_array_equals(inputOrChangeEvents, [
      'input/file-modified', 'change/file-modified',
      'input/checkbox-modified', 'change/checkbox-modified',
      'input/text-modified', 'change/text-modified',
      'input/date-modified', 'change/date-modified',
      'input/textarea-modified', 'change/textarea-modified',
      'input/select-modified', 'change/select-modified']);
}, 'Control states should be restored correctly even if controls were inserted before existing controls.');
</script>
